Rearranged some things, moving SWE-specific code to physics_swe.c.
